requirements.txt
google-generativeai>=0.3.0
markdown>=3.4.0